chore(server): create prependEntries util
