+++ lib/utils/addEntries.js
@@ -2,6 +2,7 @@
 
 const webpack = require('webpack');
 const createDomain = require('./createDomain');
+const prependEntries = require('./prependEntries');
 
 /**
  * A Entry, it can be of type string or string[] or Object<string | string[],string>
@@ -54,41 +55,6 @@ function addEntries(compiler, config, options, server) {
   } else if (options.hot) {
     hotEntry = require.resolve('webpack/hot/dev-server');
   }
-  /**
-   * getInjectedEntries Method
-   * @param {Entry} originalEntry
-   * @param {Entry} additionalEntries
-   * @returns {Entry}
-   */
-  const getInjectedEntries = (originalEntry, additionalEntries) => {
-    if (typeof originalEntry === 'function') {
-      return () =>
-        Promise.resolve(originalEntry()).then((entry) =>
-          getInjectedEntries(entry, additionalEntries)
-        );
-    }
-
-    if (typeof originalEntry === 'object' && !Array.isArray(originalEntry)) {
-      const entryObj = {};
-
-      Object.keys(originalEntry).forEach((key) => {
-        // const entryDescription = originalEntry[key];
-        // if (typeof entryDescription === 'object' && entryDescription.import) {
-        //   entryObj[key] = {
-        //     import: additionalEntries,
-        //   };
-        // } else {
-        //   entryObj[key] = additionalEntries;
-        // }
-        entryObj[key] = additionalEntries;
-      });
-
-      return entryObj;
-    }
-
-    // in this case, entry is a string or an array.
-    return additionalEntries;
-  };
 
   /**
    *
@@ -137,6 +103,7 @@ function addEntries(compiler, config, options, server) {
       additionalEntries.push(hotEntry);
     }
 
+    // add the HMR plugin to each compiler config if hot is enabled
     if (options.hot || options.hot === 'only') {
       config.plugins = config.plugins || [];
       if (
@@ -150,14 +117,16 @@ function addEntries(compiler, config, options, server) {
       }
     }
 
+    // if there are no additional entries to be added, nothing
+    // below is needed
     if (!additionalEntries.length) {
       return;
     }
 
     if (webpack.version[0] === '5') {
       // EntryPlugin is only available for webpack@5
-      // eslint-disable-next-line import/no-unresolved
-      const EntryPlugin = require('webpack/lib/EntryPlugin');
+      // @ts-ignore
+      const EntryPlugin = require('webpack/lib/EntryPlugin'); // eslint-disable-line import/no-unresolved
       // for webpack@5, we use a plugin to add global entries
       compiler.hooks.make.tapAsync(
         {
@@ -195,10 +164,7 @@ function addEntries(compiler, config, options, server) {
       // this is a hacky method of injecting entries after compiler creation
       // that should only be used for webpack@4
       const originalEntry = config.entry || './src';
-      const injectedEntries = getInjectedEntries(
-        originalEntry,
-        additionalEntries
-      );
+      const injectedEntries = prependEntries(originalEntry, additionalEntries);
 
       const compilers = compiler.compilers || [compiler];
       compilers.forEach((comp) => {
